requirements.txt
ultralytics
opencv-python
matplotlib
Pillow
PyYAML
